refactor: folder structure.
