--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "react-three-panorama"]
-	path = react-three-panorama
-	url = https://github.com/tsengyushiang/react-three-panorama.git
