--- components/Editor/index.js
@@ -1,7 +1,0 @@
-import Layout from "./Layout";
-
-const Editor = (props) => {
-  return <Layout {...props} />;
-};
-
-export default Editor;
