+++ package.json
@@ -2,8 +2,8 @@
   "name": "PanoToMesh",
   "version": "0.1.0",
   "scripts": {
-    "open-browser-mac": "open http://localhost:3000/editor/layout?image=/images/panoramas/1.jpeg",
-    "open-browser-windows": "start http://localhost:3000/editor/layout?image=/images/panoramas/1.jpeg",
+    "open-browser-mac": "open http://localhost:3000/?image=/images/panoramas/1.jpeg",
+    "open-browser-windows": "start http://localhost:3000/?image=/images/panoramas/1.jpeg",
     "open-browser": "yarn open-browser-mac || yarn open-browser-windows",
     "dev": "yarn open-browser && next dev",
     "build": "next build",
