+++ pages/index.js
@@ -1,3 +1,13 @@
-export default function Home() {
-  return null;
+import React from "react";
+import { useRouter } from "next/router";
+import Editor from "../apps";
+
+const Apps = ()=>{
+  const router = useRouter();
+
+  const image = router.query.image;
+
+  return <>{image && <Editor src={image} />}</>;
 }
+
+export default Apps;
